EZP-32164: Implemented IconPathResolver service
