+++ composer.json
@@ -13,7 +13,8 @@
         "psr-4": {
             "EzSystems\\EzPlatformCoreBundle\\": "src/EzPlatformCoreBundle/bundle/",
             "EzSystems\\EzPlatformEncoreBundle\\": "src/EzPlatformEncoreBundle/bundle/",
-            "Ibexa\\Platform\\Bundle\\Assets\\": "src/IbexaPlatformAssetsBundle/bundle/"
+            "Ibexa\\Platform\\Bundle\\Assets\\": "src/IbexaPlatformAssetsBundle/bundle/",
+            "Ibexa\\Platform\\Assets\\": "src/IbexaPlatformAssetsBundle/lib/"
         }
     },
     "autoload-dev": {
